Fix a few sign in flow issues
Attempting to load the dashboard page when signed in with an account without an Airtable record linked redirects the user to to the onboard page. The account linking step on the onboard page also requires that the email address of the currently signed-in user matches the email address of the existing account they are attempting to link, which is easier to manage then sending an email verification code.